MY TESTS PASS CONSISTENTLY!!!
Turns out it was a problem with the way my tests were written... (and not nils in the data file).  Changed random sampling of instance from array to validating each isntance of the array is an instance of the appropriate data class.  Also started reduce in revenue method from 0.0 so it always returns a float (durr).  My tests are now shorter AND test more, so win-win (plus they pass! :D)